+++ components/camel-kafka/src/test/java/org/apache/camel/component/kafka/integration/KafkaConsumerLastRecordHeaderIT.java
@@ -14,15 +14,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.kafka;
+package org.apache.camel.component.kafka.integration;
 
 import java.util.Collections;
+import java.util.List;
 import java.util.Properties;
-import java.util.stream.StreamSupport;
 
-import org.apache.camel.Endpoint;
 import org.apache.camel.EndpointInject;
+import org.apache.camel.Exchange;
 import org.apache.camel.builder.RouteBuilder;
+import org.apache.camel.component.kafka.KafkaConstants;
 import org.apache.camel.component.mock.MockEndpoint;
 import org.apache.kafka.clients.producer.ProducerRecord;
 import org.junit.jupiter.api.AfterEach;
@@ -30,18 +31,13 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
-public class KafkaConsumerTopicIsPatternTest extends BaseEmbeddedKafkaTest {
-
-    public static final String TOPIC = "test";
-    public static final String TOPIC_PATTERN = "t\\w\\wt";
-
-    @EndpointInject("kafka:" + TOPIC_PATTERN + "?topicIsPattern=true&groupId=group1&autoOffsetReset=earliest"
-                    + "&autoCommitIntervalMs=1000&sessionTimeoutMs=30000&autoCommitEnable=true&interceptorClasses=org.apache.camel.component.kafka.MockConsumerInterceptor")
-    private Endpoint from;
+public class KafkaConsumerLastRecordHeaderIT extends BaseEmbeddedKafkaTestSupport {
+    private static final String TOPIC = "last-record";
 
     @EndpointInject("mock:result")
-    private MockEndpoint to;
+    private MockEndpoint result;
 
     private org.apache.kafka.clients.producer.KafkaProducer<String, String> producer;
 
@@ -49,7 +45,6 @@ public class KafkaConsumerTopicIsPatternTest extends BaseEmbeddedKafkaTest {
     public void before() {
         Properties props = getDefaultProperties();
         producer = new org.apache.kafka.clients.producer.KafkaProducer<>(props);
-
     }
 
     @AfterEach
@@ -61,36 +56,40 @@ public class KafkaConsumerTopicIsPatternTest extends BaseEmbeddedKafkaTest {
         kafkaAdminClient.deleteTopics(Collections.singletonList(TOPIC));
     }
 
+    /**
+     * When consuming data with autoCommitEnable=false Then the LAST_RECORD_BEFORE_COMMIT header must be always defined
+     * And it should be true only for the last one
+     */
+    @Test
+    public void shouldStartFromBeginningWithEmptyOffsetRepository() throws InterruptedException {
+        result.expectedMessageCount(5);
+        result.expectedBodiesReceived("message-0", "message-1", "message-2", "message-3", "message-4");
+
+        for (int i = 0; i < 5; i++) {
+            producer.send(new ProducerRecord<>(TOPIC, "1", "message-" + i));
+        }
+
+        result.assertIsSatisfied(3000);
+
+        List<Exchange> exchanges = result.getExchanges();
+        for (int i = 0; i < exchanges.size(); i++) {
+            Boolean header = exchanges.get(i).getIn().getHeader(KafkaConstants.LAST_RECORD_BEFORE_COMMIT, Boolean.class);
+            assertNotNull(header, "Header not set for #" + i);
+            assertEquals(header, i == exchanges.size() - 1, "Header invalid for #" + i);
+            // as long as the partitions count is 1 on topic:
+            header = exchanges.get(i).getIn().getHeader(KafkaConstants.LAST_POLL_RECORD, Boolean.class);
+            assertNotNull(header, "Last record header not set for #" + i);
+            assertEquals(header, i == exchanges.size() - 1, "Last record header invalid for #" + i);
+        }
+    }
+
     @Override
     protected RouteBuilder createRouteBuilder() throws Exception {
         return new RouteBuilder() {
-
             @Override
             public void configure() throws Exception {
-                from(from).to(to);
+                from("kafka:" + TOPIC + "?groupId=A&autoOffsetReset=earliest&autoCommitEnable=false").to("mock:result");
             }
         };
     }
-
-    @Test
-    public void kafkaTopicIsPattern() throws Exception {
-        to.expectedMessageCount(5);
-        to.expectedBodiesReceivedInAnyOrder("message-0", "message-1", "message-2", "message-3", "message-4");
-        to.allMessages().header(KafkaConstants.TOPIC).isEqualTo("test");
-        // The LAST_RECORD_BEFORE_COMMIT header should not be configured on any
-        // exchange because autoCommitEnable=true
-        to.expectedHeaderValuesReceivedInAnyOrder(KafkaConstants.LAST_RECORD_BEFORE_COMMIT, null, null, null, null, null);
-
-        for (int k = 0; k < 5; k++) {
-            String msg = "message-" + k;
-            ProducerRecord<String, String> data = new ProducerRecord<>(TOPIC, "1", msg);
-            producer.send(data);
-        }
-
-        to.assertIsSatisfied(3000);
-
-        assertEquals(5, StreamSupport.stream(MockConsumerInterceptor.recordsCaptured.get(0).records(TOPIC).spliterator(), false)
-                .count());
-    }
-
 }
